+++ Account.Infrastructure.Library/Repositories/WEB/WebServiceRepository.cs
@@ -37,7 +37,7 @@ namespace Account.Application.Library.Repositories.WEB
             throw new NotImplementedException();
         }
 
-        public IEnumerable<KeyValue<long>> TitleValue()
+        public IEnumerable<KeyValue<byte>> TitleValue()
         {
             throw new NotImplementedException();
         }
